Prepare for BoringSSL update.
This change tweaks things as needed so that the code will compile
against both the BoringSSL that's currently in Android and a version
from upstream. The BORINGSSL_201509 define is temporary to allow the
switch to happen without breaking the build and a followup change will
remove it.

(cherry picked from commit 9eb9295d78a4fbfb4b4916d597a5fcb6ee2d5677)

Change-Id: I3d09b5644661353723803bcbda937d34455849a5